Nu går det kolla på en specifik genre

+++ tabla66/tabla66/Controllers/NewsController.cs
@@ -19,8 +19,13 @@ namespace tabla66.Controllers
         [HttpGet]
         public ActionResult Index()
         {
-            var news = db.News.Include(n => n.Show);
-            return View(news.ToList());
+            if (ValidateUser.IsAdmin())
+            {
+                var news = db.News.Include(n => n.Show);
+                return View(news.ToList());
+            }
+            else
+                return RedirectToAction("Index", "Shows");
         }
 
         // GET: News/Details/5
@@ -42,9 +47,14 @@ namespace tabla66.Controllers
         
         public ActionResult Create()
         {
-            ViewBag.Show_id = new SelectList(db.Show, "Id", "Title");
-            return View();
-        }
+            if (ValidateUser.IsAdmin())
+            {
+                ViewBag.Show_id = new SelectList(db.Show, "Id", "Title");
+                return View();
+            }
+            else
+                return RedirectToAction("Index", "Shows");
+    }
 
         // POST: News/Create
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
@@ -53,40 +63,50 @@ namespace tabla66.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "Id,Title,Text,Imagefile,Show_id")] News news, HttpPostedFileBase Imagefile)
         {
-            string fileName = Path.GetFileNameWithoutExtension(Imagefile.FileName);
-            string extension = Path.GetExtension(Imagefile.FileName);
-            fileName = fileName + DateTime.Now.ToString("yymmssfff") + extension;
-            news.Image = "~/Img/" + fileName;
-            fileName = Path.Combine(Server.MapPath("~/Img/"), fileName);
-            Imagefile.SaveAs(fileName);
-
-            if (ModelState.IsValid)
+            if (ValidateUser.IsAdmin())
             {
-               
-                db.News.Add(news);
-                db.SaveChanges();
-                return RedirectToAction("Index");
-            }
+                string fileName = Path.GetFileNameWithoutExtension(Imagefile.FileName);
+                string extension = Path.GetExtension(Imagefile.FileName);
+                fileName = fileName + DateTime.Now.ToString("yymmssfff") + extension;
+                news.Image = "~/Img/" + fileName;
+                fileName = Path.Combine(Server.MapPath("~/Img/"), fileName);
+                Imagefile.SaveAs(fileName);
 
-            ViewBag.Show_id = new SelectList(db.Show, "Id", "Title", news.Show_id);
-            return View(news);
+                if (ModelState.IsValid)
+                {
+               
+                    db.News.Add(news);
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+
+                ViewBag.Show_id = new SelectList(db.Show, "Id", "Title", news.Show_id);
+                return View(news);
+            }
+            else
+                return RedirectToAction("Index", "Shows");
         }
 
         // GET: News/Edit/5
         public ActionResult Edit(int? id)
         {
-            if (id == null)
+            if (ValidateUser.IsAdmin())
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                News news = db.News.Find(id);
+                if (news == null)
+                {
+                    return HttpNotFound();
+                }
+                ViewBag.Show_id = new SelectList(db.Show, "Id", "Title", news.Show_id);
+                return View(news);
             }
-            News news = db.News.Find(id);
-            if (news == null)
-            {
-                return HttpNotFound();
-            }
-            ViewBag.Show_id = new SelectList(db.Show, "Id", "Title", news.Show_id);
-            return View(news);
-        }
+            else
+                return RedirectToAction("Index", "Shows");
+    }
 
         // POST: News/Edit/5
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
@@ -95,46 +115,56 @@ namespace tabla66.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "Id,Title,Image,Text,Show_id")] News news, HttpPostedFileBase Imagefile)
         {
-
-            if (ModelState.IsValid)
+            if (ValidateUser.IsAdmin())
             {
-                if ( Imagefile!=null && Imagefile.ContentLength>0)
+
+                if (ModelState.IsValid)
                 {
-                    string filepath = Path.Combine(Server.MapPath(news.Image));
-                    System.IO.File.Delete(filepath);
+                    if ( Imagefile!=null && Imagefile.ContentLength>0)
+                    {
+                        string filepath = Path.Combine(Server.MapPath(news.Image));
+                        System.IO.File.Delete(filepath);
 
-                    string file = Path.GetFileNameWithoutExtension(Imagefile.FileName);
-                    string extension = Path.GetExtension(Imagefile.FileName);
-                    file = file + DateTime.Now.ToString("yymmssfff") + extension;
-                    news.Image = "~/Img/" + file;
+                        string file = Path.GetFileNameWithoutExtension(Imagefile.FileName);
+                        string extension = Path.GetExtension(Imagefile.FileName);
+                        file = file + DateTime.Now.ToString("yymmssfff") + extension;
+                        news.Image = "~/Img/" + file;
 
-                    file = Path.Combine(Server.MapPath("~/Img/"), file);
-                    Imagefile.SaveAs(file);
+                        file = Path.Combine(Server.MapPath("~/Img/"), file);
+                        Imagefile.SaveAs(file);
                                       
-                }
+                    }
                
-                db.Entry(news).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                    db.Entry(news).State = EntityState.Modified;
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+                ViewBag.Show_id = new SelectList(db.Show, "Id", "Title", news.Show_id);
+                return View(news);
             }
-            ViewBag.Show_id = new SelectList(db.Show, "Id", "Title", news.Show_id);
-            return View(news);
+            else
+                return RedirectToAction("Index", "Shows");
         }
 
         // GET: News/Delete/5
         public ActionResult Delete(int? id)
         {
-            if (id == null)
+            if (ValidateUser.IsAdmin())
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                News news = db.News.Find(id);
+                if (news == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(news);
             }
-            News news = db.News.Find(id);
-            if (news == null)
-            {
-                return HttpNotFound();
-            }
-            return View(news);
-        }
+            else
+                return RedirectToAction("Index", "Shows");
+    }
 
         // POST: News/Delete/5
         [HttpPost, ActionName("Delete")]
